refactor: the component structure
