+++ apps/blog/src/app/components/pages/editor/components/header/index.tsx
@@ -3,12 +3,12 @@ import { Plus } from "lucide-react";
 
 import { Button } from "@ashgw/ui";
 
-import type { SortOptions as SortOptionsType } from "./SortOptions";
+import type { SortOptions as SortOptionsType } from "./components/SortOptions";
 import type { PostDetailRo } from "~/api/models/post";
 import { useAuth } from "~/app/hooks/auth";
-import { PreviewToggle } from "../preview/PreviewToggle";
-import { ProfileButton } from "./ProfileButton";
-import { SortOptions } from "./SortOptions";
+import { PreviewToggle } from "../preview/components/PreviewToggle";
+import { ProfileButton } from "./components/ProfileButton";
+import { SortOptions } from "./components/SortOptions";
 
 interface HeaderProps {
   onClick: () => void;
